rename middleware module for queryparams
